new lint: `unnecessary_dedup_by`
