+++ clippy_utils/src/sym.rs
@@ -204,6 +204,7 @@ generate! {
     cyclomatic_complexity,
     de,
     debug_struct,
+    dedup_by,
     deprecated_in_future,
     deref_mut_method,
     diagnostics,
